techsolvo/requirements.txt
amqp==5.0.6
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
audioread==2.1.9
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coverage==5.5
decorator==5.0.9
deprecation==2.1.0
Django==3.2.6
django-audiofield==0.10.0
django-celery==3.1.17
django-rest-framework==0.1.0
djangorestframework==3.12.4
eyed3==0.9.6
filetype==1.0.7
iniconfig==1.1.1
kombu==5.1.0
mutagen==1.45.1
mysqlclient==2.0.3
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-django==4.4.0
pytz==2021.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
tinytag==1.6.0
toml==0.10.2
validators==0.18.2
vine==5.0.0
wcwidth==0.2.5
